added MEAN model (#10)
