+++ .gitignore
@@ -20,3 +20,4 @@ dask-worker-space
 .coverage.*
 .tmp-test-dir-with-unique-name/
 .spyproject/
+chet_dev/
