+++ rockhound/__init__.py
@@ -3,11 +3,15 @@
 from . import version
 from .registry import data_location
 from .etopo1 import fetch_etopo1
-from .prem import fetch_prem
+from .earth_models import fetch_prem
+from .earth_models import fetch_ak135f
+from .earth_models import fetch_iasp91
+from .earth_models import fetch_mean
 from .bedmap2 import fetch_bedmap2
 from .seafloor import fetch_seafloor_age
-from .iasp91 import fetch_iasp91
-from .ak135f import fetch_ak135f
+
+# from .iasp91 import fetch_iasp91
+# from .ak135f import fetch_ak135f
 
 
 # Get the version number through versioneer
