+++ rockhound/data_to_delete/make_registry.py
@@ -14,4 +14,4 @@ import os
 
 #%%
 
-pooch.make_registry('dat', 'dat/dat.txt')
+pooch.make_registry("dat", "dat/dat.txt")
