+++ rockhound/registry.txt
@@ -3,6 +3,7 @@ ETOPO1_Bed_g_gmt4.grd.gz e5dd1e07c9039a966096d658976145b34a7006613ba32c7e898d26d
 PREM_1s.csv 71e5d8883e5ee0529c7d8f02c8e68d12c350e993bdbd106a3697c64fa3bf9faa https://ds.iris.edu/files/products/emc/data/PREM/PREM_1s.csv
 AK135F_AVG_IDV.csv 5094703afff207e52f667ff65cc9701d86820ebe22dc84f9bed290b66b41b2da http://ds.iris.edu/files/products/emc/data/AK135F/AK135F_AVG_IDV.csv
 IASP91.csv b7c5687c78f7a27977160f055d6e247bfdcbbf2a60194924f108931dff47052f https://ds.iris.edu/files/products/emc/data/IASP91/IASP91.csv
+MEAN.nc 3180fc3c1bb1c3ffc50a3cab318f144bfd87a325fed38b0988666fafea283349 http://ds.iris.edu/files/products/emc/data/MEAN/MEAN.nc
 bedmap2_tiff.zip f4bb27ce05197e9d29e4249d64a947b93aab264c3b4e6cbf49d6b339fb6c67fe https://secure.antarctica.ac.uk/data/bedmap2/bedmap2_tiff.zip
 age.3.2.nc.bz2 49003831db78f7bb044532daf31fdc49218f75876668ee0c96877f383214b827 https://www.ngdc.noaa.gov/mgg/ocean_age/data/2008/grids/age/age.3.2.nc.bz2
 age.3.6.nc.bz2 c99358e4a980be15f05df71506d9337dc760d32e5907f9526bc8135224fc1f81 https://www.ngdc.noaa.gov/mgg/ocean_age/data/2008/grids/age/age.3.6.nc.bz2
